requirements.txt
Flask==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
gunicorn==21.2.0
python-dotenv==1.1.1
langdetect==1.0.9
fpdf==1.7.2
PyMuPDF==1.23.22
PyPDF2==3.0.1
SQLAlchemy==2.0.41
argon2-cffi==23.1.0
email-validator==2.1.0.post1
cohere==5.16.1
deep-translator==1.11.4
google-cloud-texttospeech==2.15.1
protobuf==6.31.1
psycopg2==2.9.9

# Testing
pytest==8.4.1